requirements.txt
streamlit==1.32.0
requests==2.31.0
PyPDF2==3.0.1
Pillow==10.1.0
pytesseract==0.3.10
python-dotenv==1.0.0

#brew install tesseract
#brew install tesseract-lang